Refactor: Move files due to package restructuring and update Spring configuration
