+++ api-admin/src/main/java/com/admin/ApiAdminApplication.java
@@ -1,8 +1,10 @@
-package com.app.travelboard;
+package com.admin;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
+@ComponentScan(basePackages = {"com.admin", "com.storage"})
 @SpringBootApplication
 public class ApiAdminApplication {
