+++ api-admin/src/main/resources/application.yml
@@ -1,3 +1,13 @@
 spring:
   application:
     name: api-admin
+
+  profiles:
+    active: local
+
+  config:
+    import:
+      - db-config.yml
+
+server:
+  port: 8081
