+++ api-admin/src/test/java/com/admin/ApiAdminApplicationTests.java
@@ -1,4 +1,4 @@
-package com.app.travelboard;
+package com.admin;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
