+++ api-user/src/test/java/com/user/ApiUserApplicationTests.java
@@ -1,4 +1,4 @@
-package com.app.travelboard;
+package com.user;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
