+++ gradle.properties
@@ -2,7 +2,7 @@
 applicationVersion=0.0.1
 
 ### Project configs
-projectGroup=com.app
+projectGroup=com
 
 ### Project dependency versions
 javaVersion=21
