+++ storage/src/main/java/com/storage/entity/ProductLevel.java
@@ -1,4 +1,4 @@
-package com.app.travelboard.storage.entity;
+package com.storage.entity;
 
 public enum ProductLevel {
     BASIC, PREMIUM
