+++ storage/src/main/java/com/storage/repository/AccountRepository.java
@@ -1,6 +1,6 @@
-package com.app.travelboard.storage.repository;
+package com.storage.repository;
 
-import com.app.travelboard.storage.entity.Account;
+import com.storage.entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AccountRepository extends JpaRepository<Account, Long> {
