+++ storage/src/main/java/com/storage/repository/CommentRepository.java
@@ -1,6 +1,6 @@
-package com.app.travelboard.storage.repository;
+package com.storage.repository;
 
-import com.app.travelboard.storage.entity.Comment;
+import com.storage.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
