+++ storage/src/main/java/com/storage/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package com.app.travelboard.storage.repository;
+package com.storage.repository;
 
-import com.app.travelboard.storage.entity.Post;
+import com.storage.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
